update dashMetrics responsability
