+++ src/dash/DashMetrics.js
@@ -35,62 +35,9 @@ import Round10 from './utils/Round10';
 
 /**
  * @module DashMetrics
- * @param {object} config configuration passed to DashMetrics
  */
-function DashMetrics(config) {
-
-    config = config || {};
+function DashMetrics() {
     let instance;
-    let dashManifestModel = config.dashManifestModel;
-    let manifestModel = config.manifestModel;
-
-    function getPeriod(periodId) {
-        const manifest = manifestModel.getValue();
-        if (!manifest) {
-            return -1;
-        }
-        return manifest.Period_asArray[periodId];
-    }
-
-    function getBandwidthForRepresentation(representationId, periodId) {
-        let representation;
-        let period = getPeriod(periodId);
-
-        representation = findRepresentation(period, representationId);
-
-        if (representation === null) {
-            return null;
-        }
-
-        return representation.bandwidth;
-    }
-
-    /**
-     *
-     * @param {string} representationId
-     * @param {number} periodIdx
-     * @returns {*}
-     */
-    function getIndexForRepresentation(representationId, periodIdx) {
-        let period = getPeriod(periodIdx);
-
-        return findRepresentationIndex(period, representationId);
-    }
-
-    /**
-     * This method returns the current max index based on what is defined in the MPD.
-     *
-     * @param {string} bufferType - String 'audio' or 'video',
-     * @param {number} periodIdx - Make sure this is the period index not id
-     * @return {number}
-     * @memberof module:DashMetrics
-     * @instance
-     */
-    function getMaxIndexForBufferType(bufferType, periodIdx) {
-        let period = getPeriod(periodIdx);
-
-        return findMaxBufferIndex(period, bufferType);
-    }
 
     /**
      * @param {MetricsList} metrics
@@ -309,73 +256,7 @@ function DashMetrics(config) {
         return headers;
     }
 
-    function findRepresentationIndex(period, representationId) {
-        const index = findRepresentation(period, representationId, true);
-
-        if (index !== null) {
-            return index;
-        }
-
-        return -1;
-    }
-
-    function findRepresentation(period, representationId, returnIndex) {
-        let adaptationSet,
-            adaptationSetArray,
-            representation,
-            representationArray,
-            adaptationSetArrayIndex,
-            representationArrayIndex;
-
-        if (period) {
-            adaptationSetArray = period.AdaptationSet_asArray;
-            for (adaptationSetArrayIndex = 0; adaptationSetArrayIndex < adaptationSetArray.length; adaptationSetArrayIndex = adaptationSetArrayIndex + 1) {
-                adaptationSet = adaptationSetArray[adaptationSetArrayIndex];
-                representationArray = adaptationSet.Representation_asArray;
-                for (representationArrayIndex = 0; representationArrayIndex < representationArray.length; representationArrayIndex = representationArrayIndex + 1) {
-                    representation = representationArray[representationArrayIndex];
-                    if (representationId === representation.id) {
-                        if (returnIndex) {
-                            return representationArrayIndex;
-                        } else {
-                            return representation;
-                        }
-                    }
-                }
-            }
-        }
-
-        return null;
-    }
-
-    function adaptationIsType(adaptation, bufferType) {
-        return dashManifestModel.getIsTypeOf(adaptation, bufferType);
-    }
-
-    function findMaxBufferIndex(period, bufferType) {
-        let adaptationSet,
-            adaptationSetArray,
-            representationArray,
-            adaptationSetArrayIndex;
-
-        if (!period || !bufferType) return -1;
-
-        adaptationSetArray = period.AdaptationSet_asArray;
-        for (adaptationSetArrayIndex = 0; adaptationSetArrayIndex < adaptationSetArray.length; adaptationSetArrayIndex = adaptationSetArrayIndex + 1) {
-            adaptationSet = adaptationSetArray[adaptationSetArrayIndex];
-            representationArray = adaptationSet.Representation_asArray;
-            if (adaptationIsType(adaptationSet, bufferType)) {
-                return representationArray.length;
-            }
-        }
-
-        return -1;
-    }
-
     instance = {
-        getBandwidthForRepresentation: getBandwidthForRepresentation,
-        getIndexForRepresentation: getIndexForRepresentation,
-        getMaxIndexForBufferType: getMaxIndexForBufferType,
         getCurrentRepresentationSwitch: getCurrentRepresentationSwitch,
         getLatestBufferLevelVO: getLatestBufferLevelVO,
         getCurrentBufferLevel: getCurrentBufferLevel,
